+++ R/mod_read_lacytools.R
@@ -329,6 +329,8 @@ mod_read_lacytools_server <- function(id){
     ####################  Skyline  ##########################################
     #########################################################################
     
+    # Read raw Skyline data from CSV files.
+    # Isomers are renamed.
     raw_skyline_data <- reactive({
       req(correct_file_ext(), input$data_type == "Skyline data", input$skyline_input)
       purrr::map(input$skyline_input$datapath, function(datapath) {
@@ -349,7 +351,7 @@ mod_read_lacytools_server <- function(id){
       req(raw_skyline_data())
       purrr::imap(raw_skyline_data(), function(data, i) {
         tryCatch(
-          expr = transform_skyline_data(data),
+          expr = transform_skyline_data(data, i),
           missing_columns = function(c) {
             showNotification(paste("In CSV file", i, c$message), type = "error", duration = NULL)
             shinybusy::remove_modal_spinner()